Update environment.rb due to rails3 support of inherited resources gem... more info: rake -t db:migrate => uninitialized constant Rails::Railtie
[http://www.jamievandyke.com/the-inherited-resources-railties-error, http://github.com/josevalim/inherited_resources]
If you’re like me and you skip reading a gem’s README when you install the latest, you’ll be seriously sorry when you don’t for inherited_resources because \
it is Rails 3 only. The same can be said for responders. You’ll get a big fat punch in the face with this error: